Use a temporary copy to optimize register lifetime
A temporary copy allows the compiler to keep a copy of the
copied thing and reference that over and over again, allowing
it to keep the copy in registers instead of having to look
at the memory again and again. The compiler will not copy
parts of the thing it does not need.